backend/requirements.txt
aiofiles==0.7.0
alembic==1.5.8
anyio==3.2.1
asgiref==3.4.1
async-generator==1.10
attrs==22.1.0
autopep8==1.6.0
bcrypt==3.2.0
blis==0.7.8
Brotli==1.0.9
catalogue==2.0.8
certifi==2022.9.24
cffi==1.14.6
charset-normalizer==2.0.1
chart-studio==1.1.0
click==8.0.1
colorama==0.4.4
confection==0.0.2
cryptography==3.4.7
cymem==2.0.6
ecdsa==0.17.0
exceptiongroup==1.0.1
fastapi==0.66.0
future==0.18.2
greenlet==1.1.0
h11==0.12.0
idna==3.2
interchange==2021.0.3
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.0.1
methodtools==0.4.3
monotonic==1.6
murmurhash==1.0.8
neo4j==4.3.3
nltk==3.7
numpy
outcome==1.2.0
packaging==21.0
pandas
pansi==2020.7.3
pathy==0.6.2
Pillow==9.3.0
preshed==3.0.7
py2neo==2021.2.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyinstrument==4.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.18.0
python-jose==3.3.0
python-multipart==0.0.5
python-pptx==0.6.21
pytz==2021.1
regex==2022.9.13
requests==2.26.0
retrying==1.3.3
rsa==4.7.2
scikit-learn
scipy
sd-material-ui==4.5.0
selenium==4.6.0
six==1.16.0
sklearn==0.0
smart-open==5.2.1
snakeviz==2.1.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.20
srsly==2.4.4
starlette==0.14.2
tenacity==8.0.1
thinc==8.1.2
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typer==0.4.2
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.6
uvicorn==0.14.0
wasabi==0.10.1
Werkzeug==2.0.1
wirerope==0.4.3
wsproto==1.2.0
XlsxWriter==3.0.3
yappi==1.3.6
neo4j
py2neo
uvicorn
spacy